*** Error in `./exec2': double free or corruption (fasttop): 0x00007f005c0066f0 ***

+++ Server/src/PlayerServer.hpp
@@ -11,8 +11,8 @@ int id;
 
 
 public:
-  PlayerServer(int *pos, int i);
-  int* getPosition();
+  PlayerServer(float *pos, int i);
+  float* getPosition();
   void setPositionRotation(float *pos, float *rot);
   int getID();
   void acelerar();
